[redhat] redhat: enable F29 cross-compiler install for now
Message-id: <20190503223626.19384-2-jarod@redhat.com>
Patchwork-id: 253356
O-Subject: [RHEL8 PATCH 1/2 v3] redhat: enable F29 cross-compiler install for now
Bugzilla:
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Prior to EPEL8 existing, we can use F29's cross-compiler packages to
enable cross builds. The CKI team has been using these for quite some time
now with solid results. Once EPEL-8 is available in a few weeks, that will
still be the preferred method, so this is just a stop-gap, but it may well
be helpful to a few people.

CC: Prarit Bhargava <prarit@redhat.com>
CC: Don Zickus <dzickus@redhat.com>
CC: Major Hayden <mhayden@redhat.com>

+++ redhat/scripts/download_cross.sh
@@ -5,7 +5,7 @@
 # is nothing to do.
 #
 # In some cases, it may be necessary to download & install the latest
-# EPEL7 repo information to get to the right packages.
+# EPEL8 repo information to get to the right packages.
 #
 # Argument(s) to this script is a list of rpm names to install.  If there
 # is a value for the environment variable ARCH, we will use that, too.
@@ -26,7 +26,7 @@ fi
 
 # if we're not root, all we can do now is see what's installed
 if [ "$(whoami)" != "root" ]; then
-	echo "Checking for RHEL7 cross compile packages.  If this fails, run \"make rh-cross-download\" as root."
+	echo "Checking for RHEL8 cross compile packages.  If this fails, run \"make rh-cross-download\" as root."
 	rpm -q $@
 	if [ $? == 0 ]; then
 		echo "Compilers found."
@@ -41,15 +41,29 @@ fi
 rpm -q $@ && exit 0
 
 # install epel-release if necessary
+REPO_OPTS=""
 rpm -q epel-release >& /dev/null
 if [ $? -ne 0 ]; then
-	wget -nd -r -l1 --no-parent -A "epel-release*.rpm" http://dl.fedoraproject.org/pub/epel/7/x86_64/e/
-	rpm -ivh epel-release*.rpm
-	# clean up
-	rm -f epel-release*.rpm
+	wget -nd -r -l1 --no-parent -A "epel-release*.rpm" http://dl.fedoraproject.org/pub/epel/8/x86_64/Packages/e/
+	if [ $? -eq 0 ]; then
+		rpm -ivh epel-release*.rpm
+		# clean up
+		rm -f epel-release*.rpm
+	else
+		# This is a fall-back for the time prior to EPEL-8 existing, and while functional now,
+		# will continue to be less similar to our stock compilers over time, and should not be
+		# relied upon long-term as a viable option
+		REPOURL=http://download.devel.redhat.com/released/fedora/F-29/GOLD/Everything/x86_64/os
+		REPO_OPTS="--repofrompath=rhel8fedora,$REPOURL --enablerepo=rhel8fedora"
+		# Get the fedora-gpg-keys.  This can be removed once the EPEL8 repo is active.
+		yum --nogpgcheck ${REPO_OPTS} -y install fedora-gpg-keys
+		rpm --import /etc/pki/rpm-gpg/RPM-GPG-KEY-fedora-29-primary
+	fi
 fi
 
+
+
 # install list of rpms for cross compile
-yum -y install $@
+yum ${REPO_OPTS} -y install $@
 
 exit 0
